AVX alternative to vpcmpeqd
In a particular project g++ produces code several times faster if AVX2 is set, however I want to run the program in machines having only AVX extension. So I planned to generate AVX2, translate the AVX2 instructions to corresponding AVX instructions and link the program. With this code I test an alternative for VCMPEQD instruction, used by g++ to get the all-ones register.